refactor: organise code.
