+++ web/api/history_handler.py
@@ -9,7 +9,7 @@ from tornado.web import HTTPError
 
 from web.api.base_handlers import BaseRequestHandler
 from web.api.models import ChatHistoryModel, MessageModel
-from web.api.utils import authenticated
+from web.api.utils.pydantic_utils import authenticated
 from web.utils.streamlit_application import st_app
