+++ web/api/space_handler.py
@@ -8,7 +8,7 @@ from pydantic import BaseModel, ValidationError
 from tornado.web import HTTPError
 
 from web.api.base_handlers import BaseRagRequestHandler
-from web.api.utils import authenticated
+from web.api.utils.pydantic_utils import authenticated
 from web.utils.streamlit_application import st_app
