+++ web/api/token_handler.py
@@ -7,7 +7,7 @@ from pydantic import BaseModel, ValidationError
 
 from web.api.base_handlers import BaseRequestHandler
 from web.api.models import UserModel
-from web.api.utils import decode_jwt, encode_jwt
+from web.api.utils.auth_utils import decode_jwt, encode_jwt
 from web.utils.streamlit_application import st_app
 
 
@@ -36,10 +36,12 @@ class TokenResponseModel(BaseModel):
     expires_in: int
     refresh_token: Optional[str] = None
 
-class TokenErrorResponseModel(BaseModel):
-    """Token error response model."""
+class ToeknErrorResponseModel(BaseModel):
+    """General error response model."""
+
     error: str
     error_description: str
+
 
 
 @st_app.api_route("/api/token")
@@ -49,7 +51,7 @@ class TokenHandler(BaseRequestHandler):
     def token_error(self: Self, reason: str, status: int = 400) -> None:
         """Handle token error."""
         self.set_status(status)
-        self.write(TokenErrorResponseModel(error='invalid_request', error_description=reason).json())
+        self.write(ToeknErrorResponseModel(error="invalid_request", error_description=reason).json())
         self.finish()
 
     def post(self: Self) -> None:
