+++ web/api/top_questions_handler.py
@@ -10,7 +10,7 @@ from llama_index.indices.base import BaseIndex
 from tornado.web import HTTPError
 
 from web.api.base_handlers import BaseRagRequestHandler
-from web.api.utils import authenticated
+from web.api.utils.pydantic_utils import authenticated
 from web.utils.streamlit_application import st_app
